requirements.txt
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
openai==1.6.1
pandas==2.0.1
numpy==1.24.3
rank-bm25==0.2.2
streamlit==1.29.0
streamlit-authenticator==0.2.3
tiktoken==0.5.1
weaviate-client==4.4b6
tqdm
langchain